Bugfixes, Wood Elves hero crash hopefully fixed
Lots of misc fixes to things that people missed and didn't check. Lots
of woundfends, a couple of divine inspirations, the majority of them in
orcs. Wood Elf multiheroes deleted as they were just copystatted other
units. The original units are now recruitable and multiheroes - that
should be fine as long as it's allowed.